// packages/core/src/generated/types.ts
export interface BatchInvoiceType {
  /**
   * Sequence number of the modification document within the batch
   *
   * A módosító okirat sorszáma a kötegen belül
   */
  batchIndex: number;
  /**
   * Data of a single invoice or modification document
   *
   * Egy számla vagy módosító okirat adatai
   */
  invoice: InvoiceType;
}
// ...
export interface CustomerInfoType {
  /**
   * Customers status by VAT
   *
   * Vevő ÁFA szerinti státusza
   */
  customerVatStatus: CustomerVatStatusType;
  /**
   * VAT subjectivity data of the customer
   *
   * A vevő ÁFA alanyisági adatai
   */
  customerVatData?: CustomerVatDataType;
  /**
   * Name of the customer
   *
   * A vevő neve
   */
  customerName?: string;
  /**
   * Address of the customer
   *
   * A vevő címe
   */
  customerAddress?: AddressType;
  /**
   * Bank account number of the customer
   *
   * Vevő bankszámlaszáma
   */
  customerBankAccountNumber?: string;
}
// ...
export interface FiscalRepresentativeType {
  /**
   * Tax number of the fiscal representative
   *
   * A pénzügyi képviselő adószáma
   */
  fiscalRepresentativeTaxNumber: TaxNumberType;
  /**
   * Name of the fiscal representative
   *
   * A pénzügyi képviselő neve
   */
  fiscalRepresentativeName: string;
  /**
   * Address of the fiscal representative
   *
   * Pénzügyi képviselő címe
   */
  fiscalRepresentativeAddress: AddressType;
  /**
   * Bank account number opened by the fiscal representative for the issuer of
   * the invoice (supplier)
   *
   * Pénzügyi képviselő által a számla kibocsátó (eladó) számára megnyitott
   * bankszámla bankszámlaszáma
   */
  fiscalRepresentativeBankAccountNumber?: string;
}
// ...
export interface InvoiceData {
  /**
   * Sequential number of the original invoice or modification document - section
   * 169 (b) or section 170 (1) b) of the VAT law
   *
   * Számla vagy módosító okirat sorszáma - ÁFA tv. 169. § b) vagy 170. § (1)
   * bek. b) pont
   */
  invoiceNumber: string;
  /**
   * Date of issue of the invoice or the modification document - section 169 (a)
   * of the VAT law, section 170 (1) a) of the VAT law
   *
   * Számla vagy módosító okirat kelte - ÁFA tv. 169. § a), ÁFA tv. 170. § (1)
   * bek. a)
   */
  invoiceIssueDate: string;
  /**
   * Indicates whether the data exchange is identical with the invoice (the
   * invoice does not contain any more data)
   *
   * Jelöli, ha az adatszolgáltatás maga a számla (a számlán nem szerepel több
   * adat)
   */
  completenessIndicator: boolean;
  /**
   * A common type to describe invoice information
   *
   * Számlaadatok leírására szolgáló közös típus
   */
  invoiceMain: InvoiceMainType;
}
// ...
export interface InvoiceHeadType {
  /**
   * Data related to the issuer of the invoice (supplier)
   *
   * Számla kibocsátó (eladó) adatai
   */
  supplierInfo: SupplierInfoType;
  /**
   * Data related to the customer
   *
   * Vevő adatai
   */
  customerInfo?: CustomerInfoType;
  /**
   * Data related to the fiscal representative
   *
   * Pénzügyi képviselő adatai
   */
  fiscalRepresentativeInfo?: FiscalRepresentativeType;
  /**
   * Invoice detail adata
   *
   * Számla részletező adatok
   */
  invoiceDetail: InvoiceDetailType;
}
// ...
export interface InvoiceMainType {
  /**
   * Data of a single invoice or modification document
   *
   * Egy számla vagy módosító okirat adatai
   */
  invoice?: InvoiceType;
  /**
   * Data of a batch of modification documents
   *
   * Kötegelt módosító okirat adatai
   */
  batchInvoice?: BatchInvoiceType[];
}
// ...
export interface InvoiceReferenceType {
  /**
   * Sequence number of the original invoice, on which the modification occurs -
   * section 170 (1) c) of the VAT law
   *
   * Az eredeti számla sorszáma, melyre a módosítás vonatkozik - ÁFA tv. 170. §
   * (1) c)
   */
  originalInvoiceNumber: string;
  /**
   * Indicates whether the modification references to an original invoice which
   * is not and will not be exchanged
   *
   * Annak jelzése, hogy a módosítás olyan alapszámlára hivatkozik, amelyről nem
   * történt és nem is fog történni adatszolgáltatás
   */
  modifyWithoutMaster: boolean;
  /**
   * The unique sequence number referring to the original invoice
   *
   * A számlára vonatkozó módosító okirat egyedi sorszáma
   */
  modificationIndex: number;
}

/**
 * Processing status of the invoice
 *
 * A számla feldolgozási státusza
 */
export type InvoiceStatusType = 'RECEIVED' | 'PROCESSING' | 'SAVED' | 'DONE' | 'ABORTED';

/**
 * Data of a single invoice or modification document
 *
 * Egy számla vagy módosító okirat adatai
 */
export interface InvoiceType {
  /**
   * Modification or cancellation data
   *
   * A módosítás vagy érvénytelenítés adatai
   */
  invoiceReference?: InvoiceReferenceType;
  /**
   * Data concerning the whole invoice
   *
   * A számla egészét jellemző adatok
   */
  invoiceHead: InvoiceHeadType;
  /**
   * Product/service data appearing on the invoice
   *
   * A számlán szereplő tételek adatai
   */
  invoiceLines?: LinesType;
  /**
   * Summary data of product charges
   *
   * Termékdíjjal kapcsolatos összesítő adatok
   *
   * @remarks At most 2 entries.
   */
  productFeeSummary?: ProductFeeSummaryType[];
  /**
   * Summary data according to VAT law
   *
   * Az ÁFA törvény szerinti összesítő adatok
   */
  invoiceSummary: SummaryType;
}
// ...
export interface LineAmountsNormalType {
  /**
   * Line net data
   *
   * Tétel nettó adatok
   */
  lineNetAmountData: LineNetAmountDataType;
  /**
   * Tax rate or tax exemption marking
   *
   * Adómérték vagy adómentesség jelölése
   */
  lineVatRate: VatRateType;
  /**
   * Line VAT data
   *
   * Tétel ÁFA adatok
   */
  lineVatData?: LineVatDataType;
  /**
   * Line gross data
   *
   * Tétel bruttó adatok
   */
  lineGrossAmountData?: LineGrossAmountDataType;
}
// ...
export interface LinesType {
  /**
   * Indicates whether the data exchange contains merged line data due to size
   * reduction
   *
   * Jelöli, ha az adatszolgáltatás méretcsökkentés miatt összevont soradatokat
   * tartalmaz
   */
  mergedItemIndicator: boolean;
  /**
   * Product / service item
   *
   * Termék/szolgáltatás tétel
   */
  line: LineType[];
}

/**
 * Field type including data of invoice items (product or service)
 *
 * A számla tételek (termék vagy szolgáltatás) adatait tartalmazó típus
 *
 * @remarks At most one of `lineAmountsNormal`, `lineAmountsSimplified` may be set.
 */
export interface LineType {
// ...
  lineAmountsNormal?: LineAmountsNormalType;
// ...
}
// ...
export interface PaymentEvidenceDocumentDataType {
  /**
   * Sequential number of the invoice, or other document considered as such
   *
   * Számla sorszáma vagy egyéb okirat azonosító száma
   */
  evidenceDocumentNo: string;
  /**
   * Date of issue of the invoice
   *
   * Számla kelte
   */
  evidenceDocumentDate: string;
  /**
   * Name of obligator
   *
   * Kötelezett neve
   */
  obligatedName: string;
  /**
   * Address of obligator
   *
   * Kötelezett címe
   */
  obligatedAddress: AddressType;
  /**
   * Tax number of obligated
   *
   * A kötelezett adószáma
   */
  obligatedTaxNumber: TaxNumberType;
}
// ...
export interface ProductFeeDataType {
  /**
   * Product charges code (Kt or Csk code)
   *
   * Termékdíj kód (Kt vagy Csk)
   */
  productFeeCode: ProductCodeType;
  /**
   * Quantity of product, according to product charge
   *
   * A termékdíjjal érintett termék mennyisége
   */
  productFeeQuantity: string;
  /**
   * Unit of the rate (kg or piece)
   *
   * A díjtétel egysége (kg vagy darab)
   */
  productFeeMeasuringUnit: ProductFeeMeasuringUnitType;
  /**
   * Product fee rate (HUF/unit)
   *
   * A termékdíj díjtétele (HUF/egység)
   */
  productFeeRate: string;
  /**
   * Amount in Hungarian forints of the product fee
   *
   * Termékdíj összege forintban
   */
  productFeeAmount: string;
}

/**
 * Unit of the rate type
 *
 * Díjtétel egység típus
 */
export type ProductFeeMeasuringUnitType = 'DARAB' | 'KG';

/**
 * Product fee summary type
 *
 * Termékdíj összesítés típus
 */
export type ProductFeeOperationType = 'REFUND' | 'DEPOSIT';

/**
 * Summary of product charges
 *
 * Termékdíj összegzés adatok
 */
export interface ProductFeeSummaryType {
  /**
   * Indicating whether the the product fee summary concerns refund or deposit
   *
   * Annak jelzése, hogy a termékdíj összesítés visszaigénylésre (REFUND) vagy
   * raktárba történő beszállításra (DEPOSIT) vonatkozik
   */
  productFeeOperation: ProductFeeOperationType;
  /**
   * Product charges data
   *
   * Termékdíj adatok
   */
  productFeeData: ProductFeeDataType[];
  /**
   * Aggregate product charges
   *
   * Termékdíj összesen
   */
  productChargeSum: string;
  /**
   * Data of the document verifying the declaration submitted on the product fee
   * according to the Paragraph (3), Section 13 and the Paragraph (3) Section 25
   * of the Act LXXXV of 2011
   *
   * A termékdíj bevallását igazoló dokumentum adatai a 2011. évi LXXXV. tv. 13.
   * § (3) szerint és a 25. § (3) szerint
   */
  paymentEvidenceDocumentData?: PaymentEvidenceDocumentDataType;
}
// ...
export interface SummaryGrossDataType {
  /**
   * Gross amount of the invoice expressed in the currency of the invoice
   *
   * A számla bruttó összege a számla pénznemében
   */
  invoiceGrossAmount: string;
  /**
   * Gross amount of the invoice expressed in HUF
   *
   * A számla bruttó összege forintban
   */
  invoiceGrossAmountHUF: string;
}

/**
 * Calculation of invoice totals (not simplified invoice)
 *
 * Számla összesítés (nem egyszerűsített számla esetén)
 */
export interface SummaryNormalType {
  /**
   * Calculation of invoice totals per VAT rates
   *
   * Összesítés ÁFA-mérték szerint
   */
  summaryByVatRate: SummaryByVatRateType[];
  /**
   * Net amount of the invoice expressed in the currency of the invoice
   *
   * A számla nettó összege a számla pénznemében
   */
  invoiceNetAmount: string;
  /**
   * Net amount of the invoice expressed in HUF
   *
   * A számla nettó összege forintban
   */
  invoiceNetAmountHUF: string;
  /**
   * VAT amount of the invoice expressed in the currency of the invoice
   *
   * A számla ÁFA összege a számla pénznemében
   */
  invoiceVatAmount: string;
  /**
   * VAT amount of the invoice expressed in HUF
   *
   * A számla ÁFA összege forintban
   */
  invoiceVatAmountHUF: string;
}

/**
 * Calculation of simplified invoice totals
 *
 * Egyszerűsített számla összesítés
 */
export interface SummarySimplifiedType {
  /**
   * Marking the tax rate or the fact of tax exemption
   *
   * Adómérték vagy adómentesség jelölése
   */
  vatRate: VatRateType;
  /**
   * The gross amount of the sale or service for the given tax amount in the
   * currency of the invoice
   *
   * Az adott adótartalomhoz tartozó értékesítés vagy szolgáltatásnyújtás bruttó
   * összege a számla pénznemében
   */
  vatContentGrossAmount: string;
  /**
   * The gross amount of the sale or service for the given tax amount in HUF
   *
   * Az adott adótartalomhoz tartozó értékesítés vagy szolgáltatásnyújtás bruttó
   * összege forintban
   */
  vatContentGrossAmountHUF: string;
}

/**
 * Data of calculation of invoice totals
 *
 * Számla összesítés adatai
 *
 * @remarks Exactly one of `summaryNormal`, `summarySimplified` may be set.
 */
export interface SummaryType {
  /**
   * Calculation of invoice totals (not simplified invoice)
   *
   * Számla összesítés (nem egyszerűsített számla esetén)
   */
  summaryNormal?: SummaryNormalType;
  /**
   * Calculation of simplified invoice totals
   *
   * Egyszerűsített számla összesítés
   */
  summarySimplified?: SummarySimplifiedType[];
  /**
   * Gross data of the invoice summary
   *
   * A számla összesítő bruttó adatai
   */
  summaryGrossData?: SummaryGrossDataType;
}

/**
 * Company codes of the supplier
 *
 * Az eladó vállalati kódjai
 */
export interface SupplierCompanyCodesType {
  /**
   * Company code of the supplier
   *
   * Az eladó vállalati kódja
   */
  supplierCompanyCode: string[];
}

/**
 * Invoice supplier (seller) data
 *
 * A szállító (eladó) adatai
 */
export interface SupplierInfoType {
  /**
   * Tax number or group identification number
   *
   * Belföldi adószám vagy csoportazonosító szám
   */
  supplierTaxNumber: TaxNumberType;
  /**
   * Tax number of group member, when the supply of goods or services is done
   * under group identification number
   *
   * Csoport tag adószáma, ha a termékbeszerzés vagy szolgáltatás nyújtása
   * csoportazonosító szám alatt történt
   */
  groupMemberTaxNumber?: TaxNumberType;
  /**
   * Community tax number
   *
   * Közösségi adószám
   */
  communityVatNumber?: string;
  /**
   * Name of the seller (supplier)
   *
   * Az eladó (szállító) neve
   */
  supplierName: string;
  /**
   * Address of the seller (supplier)
   *
   * Az eladó (szállító) címe
   */
  supplierAddress: AddressType;
  /**
   * Bank account number of the seller (supplier)
   *
   * Az eladó (szállító) bankszámlaszáma
   */
  supplierBankAccountNumber?: string;
  /**
   * Value is true if the seller (supplier) is individually exempted from VAT
   *
   * Értéke true, amennyiben az eladó (szállító) alanyi ÁFA mentes
   */
  individualExemption?: boolean;
  /**
   * Number of supplier’s tax warehouse license or excise license (Act LXVIII of
   * 2016)
   *
   * Az eladó adóraktári engedélyének vagy jövedéki engedélyének száma (2016. évi
   * LXVIII. tv.)
   */
  exciseLicenceNum?: string;
}